move our Section component in file structure
