+++ website/src/components/brand-picker/brand-picker.js
@@ -5,7 +5,7 @@ import { Fragment } from 'react';
 import { useBrandSwitcher } from '../providers/brand-switcher';
 import { useRouter } from 'next/router';
 import { Container, Grid, Cell } from '@westpac/grid';
-import { Section } from '../../components/layout/section';
+import { Section } from '../../components/section';
 import { Body } from '../../components/body';
 import { BASE_URL } from '../../config.js';
 import { findByDisplayValue } from '@testing-library/react';
