+++ website/src/components/layout/index.js
@@ -3,5 +3,4 @@ export { Gridly } from './gridly';
 export { Layout } from './layout';
 export { Navigation } from './navigation';
 export { Normalize } from './normalize';
-export { Section } from './section';
 export { Sidebar } from './sidebar';
