+++ website/src/components/section/index.js
@@ -1,0 +1,2 @@
+export { Section } from './section';
+export { SectionHeading } from './section-heading';
