+++ website/src/pages/design-system/index.js
@@ -6,7 +6,7 @@ import { Button } from '@westpac/button';
 import { Heading } from '@westpac/heading';
 import { TextInput } from '@westpac/text-input';
 import HomePageHeader from '../../components/header/home-page-header';
-import { Section } from '../../components/layout';
+import { Section } from '../../components/section';
 import { Body } from '../../components/body';
 import { BlockList, BlockListItem as Item } from '../../components/block-list';
 import { Footer } from '../../components/layout/footer';
